requirements-dev.txt
-r requirements.txt
flake8
pylint
pylint-django
pytest
pytest-django
sphinx
coverage
python_coveralls
django-debug-toolbar
sqlparse>=0.4.2 # not directly required, pinned by Snyk to avoid a vulnerability
